+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Added
 - `proseg` support for Visium HD data (see the related tutorial).
+- Bypass spatial join when using only one patch to speed-up `make_transcript_patches` (#362)
 
 ## [2.1.9] - 2025-11-14